combine lib/workspace and lib/workspaceSocket
